Unstable dev

See merge request anywave/AnyWave!6

+++ src/AnyWave/Display/AwDisplay.cpp
@@ -116,7 +116,7 @@ AwChannelList AwDisplay::displayedChannels()
 AwSignalView *AwDisplay::addSignalView(AwViewSetup *setup)
 {
 	// Set other views flag
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->setProcessFlags(AwSignalView::NoProcessUpdate);
 
 	AwSignalView *view = new AwSignalView((AwViewSettings *)setup);
@@ -127,7 +127,7 @@ AwSignalView *AwDisplay::addSignalView(AwViewSetup *setup)
 	QList<AwDisplayPlugin *> plugins = AwPluginManager::getInstance()->displays();
 	QList<AwProcessPlugin *> QTSplugins = pm->processPluginsWithFeatures(Aw::ProcessFlags::PluginAcceptsTimeSelections);
 
-	foreach (AwDisplayPlugin *plugin, plugins)
+	for (auto plugin : plugins)
 		view->addNewDisplayPlugin(plugin);
 
 	m_signalViews << view;
@@ -157,21 +157,20 @@ AwSignalView *AwDisplay::addSignalView(AwViewSetup *setup)
 	if (!m_virtualChannels.isEmpty())
 		view->addVirtualChannels(m_virtualChannels);
 
-//	m_splitterWidget->addWidget(view->widget());
 	m_splitterWidget->addWidget(view);
 	m_splitterWidget->repaint();
 	
 	// set flags so that the views inform the Process Manager about changes.
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->setProcessFlags(AwSignalView::UpdateProcess);
 
 	view->setMarkers(AwMarkerManager::instance()->displayedMarkers());
 	
 	// QTS
 	QStringList list;
-	foreach(AwProcessPlugin *p, QTSplugins) {
+	for (auto p : QTSplugins) 
 		list << p->name;
-	}
+	
 	view->scene()->setQTSPlugins(list);
 	connect(view->scene(), SIGNAL(processSelectedForLaunch(QString&, AwChannelList&, float, float)),
 		pm, SLOT(launchQTSPlugin(QString&, AwChannelList&, float, float)));
@@ -209,7 +208,7 @@ void AwDisplay::closeFile()
 	m_virtualChannels.clear();
 	AwDisplaySetupManager::instance()->resetToDefault();
 	addMarkerModeChanged(false);
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->closeFile();
 	m_allSelectedChannels.clear();
 }
@@ -223,7 +222,10 @@ void AwDisplay::quit()
 
 void AwDisplay::saveChannelSelections()
 {
-	QString path = AwSettings::getInstance()->filePath() + ".sel";
+	auto fi = AwSettings::getInstance()->fileInfo();
+	if (!fi)
+		return;
+	QString path = fi->filePath() + ".sel";
 	if (m_selectedLabels.isEmpty()) {
 		if (QFile::exists(path))
 			QFile::remove(path);
@@ -232,7 +234,7 @@ void AwDisplay::saveChannelSelections()
 	QFile file(path);
 	if (file.open(QIODevice::ReadWrite|QIODevice::Text|QIODevice::Truncate)) {
 		QTextStream stream(&file);
-		foreach(QString label, m_selectedLabels)
+		for (auto label : m_selectedLabels)
 			stream << label << endl;
 		file.close();
 	}
@@ -240,7 +242,7 @@ void AwDisplay::saveChannelSelections()
 
 void AwDisplay::loadChannelSelections()
 {
-	QString path = AwSettings::getInstance()->filePath() + ".sel";
+	QString path = AwSettings::getInstance()->fileInfo()->filePath() + ".sel";
 	QFile file(path);
 	QStringList labels;
 	if (file.open(QIODevice::ReadOnly|QIODevice::Text)) {
@@ -303,7 +305,7 @@ void AwDisplay::captureViews()
 	// first, clear the lastCaptureFile set in Settings
 	aws->lastCaptureFile.clear();  // so in case of error when capturing, the file remains empty.
 
-	QString dir = aws->currentFileDir();
+	QString dir = aws->fileInfo()->dirPath();
 	int count = 1;
 	QString filename = QString("capture_%1.png").arg(count++);
 	QString file = QString("%1/%2").arg(dir).arg(filename);
@@ -474,7 +476,7 @@ void AwDisplay::alignViewsHorizontaly()
 
 void AwDisplay::setSelectedChannelsFromLabels(const QStringList& labels)
 {
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->selectChannelsFromLabels(labels);
 }
 
@@ -498,7 +500,7 @@ void AwDisplay::updateICAChannelRejected(const QString& label, bool rejected)
 
 void AwDisplay::makeChannelVisible(const QString& label)
 {
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->scene()->showChannel(label);
 }
 
@@ -508,7 +510,7 @@ void AwDisplay::makeChannelVisible(const QString& label)
 /// check for all channels corresponding with name parameter and set their selected state to parameter selected.
 void AwDisplay::changeChannelsSelectionState(const QString& name, bool selected)
 {
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->changeChannelSelectionState(name, selected);
 }
 
@@ -522,19 +524,19 @@ void AwDisplay::addVirtualChannels(AwChannelList *channels)
 
 	m_virtualChannels += *channels;
 	// add the channels all signal views.
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->addVirtualChannels(*channels);
 }
 
 void AwDisplay::removeICAChannels()
 {
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v: m_signalViews)
 		v->removeICAChannels();
 
 	// get channel list back from scenes
 	AwChannelList completeList;
 
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		completeList += v->displayedChannels();
 
 	// send new displayed channels event
@@ -547,10 +549,10 @@ void AwDisplay::removeVirtualChannels(AwChannelList *channels)
 	if (channels->isEmpty())
 		return;
 
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->removeVirtualChannels(*channels);
 
-	foreach(AwChannel *c, *channels)
+	for (auto c : *channels)
 		if (m_virtualChannels.contains(c))
 			m_virtualChannels.removeAll(c);
 }
@@ -558,7 +560,7 @@ void AwDisplay::removeVirtualChannels(AwChannelList *channels)
 
 void AwDisplay::updateDisplay()
 {
-	foreach (AwSignalView *v, m_signalViews)
+	for (auto v : m_signalViews)
 		v->update();
 }
 
@@ -622,7 +624,6 @@ void AwDisplay::changeCurrentSetup(AwDisplaySetup *newSetup)
 
 	// remove all the current views
 	foreach (AwSignalView *v, m_signalViews) {
-//		v->widget()->setParent(NULL);
 		v->setParent(NULL);
 		delete v;
 	}
